requirements.txt
Django==5.0.2
openpyxl==3.1.2
pandas==2.2.1
PyYAML==6.0.1
python-decouple==3.8
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
django-extensions==3.2.3
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0 